+++ src/components/functional/sign-transactions-panel/components/sign-transactions-advanced/components/sign-transactions-advanced-data/components/sign-transactions-advanced-data-decode/sign-transactions-advanced-data-decode.tsx
@@ -1,5 +1,6 @@
 import { Component, h, Prop } from '@stencil/core';
 import classNames from 'classnames';
+import { Icon } from 'common/Icon';
 import { DecodeMethodEnum } from 'components/functional/sign-transactions-panel/sign-transactions-panel.types';
 import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 
@@ -30,7 +31,8 @@ export class SignTransactionsAdvancedDataDecode {
           {this.currentDecodeMethod}
         </div>
 
-        <mvx-single-angle-down-icon
+        <Icon
+          name="angle-down"
           class={classNames('sign-transactions-advanced-data-decode-icon', {
             [signTransactionsAdvancedDataDecodeClasses.icon]: true,
             [signTransactionsAdvancedDataDecodeClasses.iconRotated]: this.isToggled,