+++ src/components/functional/sign-transactions-panel/components/sign-transactions-footer/sign-transactions-footer.tsx
@@ -1,5 +1,6 @@
 import { Component, Fragment, h, State } from '@stencil/core';
 import classNames from 'classnames';
+import { Icon } from 'common/Icon';
 import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 
 import state from '../../signTransactionsPanelStore';
@@ -131,7 +132,7 @@ export class SignTransactionsFooter {
                     </span>
                   ) : (
                     <span class="sign-transactions-footer-button-icon">
-                      {needsSigning ? <mvx-pencil-icon /> : <mvx-check-icon />}
+                      <Icon name={!needsSigning ? 'pencil' : 'check'} />
                     </span>
                   )}
                 </span>
@@ -167,12 +168,7 @@ export class SignTransactionsFooter {
             />
           )}
 
-          <mvx-copy-button
-            text={username ?? address}
-            class="sign-transactions-footer-identity-copy"
-            iconClass="sign-transactions-footer-identity-copy-icon"
-          />
-
+          <mvx-copy-button text={username ?? address} class="sign-transactions-footer-identity-copy" />
           <mvx-explorer-link link={explorerLink} iconClass={signTransactionsFooterClasses.explorerLinkIcon} />
         </div>
       </div>